requirements.txt
feedparser>=6.0.0
yt-dlp>=2024.0.0
google-genai>=1.0.0
requests>=2.31.0
